% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myconstr.R
\name{myttest}
\alias{myttest}
\title{A constructor function named myttest is used for conducting t-test}
\usage{
myttest(x, y, alpha = 0.05, paired = FALSE)
}
\arguments{
\item{x}{vector of data}

\item{y}{vector of data}

\item{alpha}{alpha level between 0,1}

\item{paired}{argument that is TRUE when data is paired and FALSE otherwise}
}
\value{
The constructor function will return a list containing the following
information in a list.

1. Which Type of t-test was conducted depending on the properties of the
  input data. The available options for different t-tests include "Welch",
  "T-Test"  or "Paired" t-test. When the input data x and y are paired, then
  we conduct the "Paired" t-test. For non- paired data, if the underlying
  variances are  the same(equal) , then "T-test" t-test is conducted. For
  non- paired data, if the underlying variances are he different(unequal),
  then "Welch" t-test is conducted.

2. A Y/N (Y= Yes, N = No) conclusion to whether we reject the null hypothesis
  or not at the alpha level of significance

3. The summary statistics obtained from the appropriate t.test.

4. The data in a list (data vectors x and y ).

This list is then released invisibly with the "Rttest" class.
}
\description{
This constructor function will take two vectors x and y. It will
check whether the data is paired or not.

 If the data is paired, then , we run  the relevant t-test, where argument
 paired is set to TRUE. If the data is not paired ,an f-test is conducted to
 check for equality of variances. Subsequently, the appropriate t-test is
 conducted depending the variances are equal or not. After conducting the
 appropriate t-test, a list a returned that shows whether, "Welch", "T-Test"
 or "Paired" t-test was conducted, a Y/N conclusion as to whether we should
 reject the NULL of equality of means at the alpha level of significance,
 summary statistics made by the appropriate t.test, The data in a list.

 Then, we we release the above list invisibly with the class "Rttest".
}
\examples{
myttest(x=rnorm(30,10,12), y=rnorm(40, 7, 10))

}
