% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPlot.R
\name{plot.Rttest}
\alias{plot.Rttest}
\title{Method for plotting the class Rttest}
\usage{
\method{plot}{Rttest}(x, ...)
}
\arguments{
\item{x}{An object of class Rttest which comes from the t test}

\item{...}{Argument that will be passed to print.Rttest(unused currently)}
}
\value{
A boxplot of the two samples using \code{ggplot()}
}
\description{
plot.Rttest is a function to streamlining the process of visualizing t-test
results by creating appropriate and well labelled box plots for both paired
and non- paired data. The function will generate appropriate box plots for
input data. If the data is non - paired, then, the plot.Rttest function will
 provide side by side box plots of x and y.

 On the other hand. if the data is paired, then, the function will make a
 box plot of differences. Furthermore, the confidence interval  for the
 difference of means is plotted inside the box plot of differences. The plot
 function also ensures that all the box plots are well labelled.

 If the input data is non- paired, then, function will return side by side
 box plots of x and y, with a title "Box plots for x and y".

For paired input data, this function will return a box plot of differences,
titled "Difference between x and y" with the confidence interval for the
difference of means plotted inside the box plot of differences.
}
\details{
The function \code{myttest()} will produce a list of class Rttest, which
consists of three components. These are sourced from the object and in
\code{plot.Rttestt()}
}
\examples{
l <- myttest(x=rnorm(30), y=rnorm(40,0.5));plot(l)

}
