% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPrint.R
\name{print.Rttest}
\alias{print.Rttest}
\title{Method for printing the class Rttest}
\usage{
\method{print}{Rttest}(x, ...)
}
\arguments{
\item{x}{An object of class Rttest which comes from the t-test}

\item{...}{Argument that will be passed to print.Rttest, and extra options to
be sent to method}
}
\value{
The print.Rttest function  will print A(1-alpha)*100 confidence
interval for μₓ -  μᵧ.
}
\description{
The print.Rttest function will take an object of the class "Rttest" and,
present the  t-test results in a comprehensive and intuitive way.The print
function will print A(1-alpha)*100 confidence interval for μₓ -  μᵧ.
}
\details{
The function \code{myttest()} will produce a list of class Rttest, which
consists of three components. These are sourced from the object and in
\code{print.Rttestt()}
}
\examples{
l <- myttest(x=rnorm(30), y=rnorm(40,0.5))
print(l)



}
